rg: refactored resultset classes, support multiple result series...
